[gcp] consolidate gcp helpers into base helpers
