+++ gcp_audit_rules/gcp_iam_admin_role_assigned.py
@@ -1,5 +1,5 @@
 from fnmatch import fnmatch
-from panther_gcp_helpers import get_binding_deltas
+from panther_base_helpers import get_binding_deltas
 from panther_base_helpers import deep_get
 
 ADMIN_ROLES = {
