--- global_helpers/panther_gcp_helpers.yml
@@ -1,5 +1,0 @@
-AnalysisType: global
-Filename: panther_gcp_helpers.py
-GlobalID: panther_gcp_helpers
-Description: >
-  Used to define global helpers and variables.
